+++ run.sh
@@ -68,17 +68,17 @@ then
 fi
 
 # Database user
-if [[ -n "$dbEngine" && ${dbEngine} == "MySQL" ]]
-then
-    echo -en "\n"
-    echo "${RED}Create the user account that will be allowed to access this database and flush the privileges to notify the MySQL server of the changes${RESET}"
-    if [[ ${localOs} == "Windows" ]]
-    then
-      winpty docker container exec -it $PROJECT_NAME-mysql mysql -u$MYSQL_ROOT_USER -p$MYSQL_ROOT_PASSWORD -e "GRANT ALL ON $MYSQL_DATABASE.* TO '$MYSQL_USER'@'%' IDENTIFIED BY '$MYSQL_PASSWORD';FLUSH PRIVILEGES;"
-    else
-      docker container exec -it $PROJECT_NAME-mysql mysql -u$MYSQL_ROOT_USER -p$MYSQL_ROOT_PASSWORD -e "GRANT ALL ON $MYSQL_DATABASE.* TO '$MYSQL_USER'@'%' IDENTIFIED BY '$MYSQL_PASSWORD';FLUSH PRIVILEGES;"
-    fi
-fi
+#if [[ -n "$dbEngine" && ${dbEngine} == "MySQL" ]]
+#then
+#    echo -en "\n"
+#    echo "${RED}Create the user account that will be allowed to access this database and flush the privileges to notify the MySQL server of the changes${RESET}"
+#    if [[ ${localOs} == "Windows" ]]
+#    then
+#      winpty docker container exec -it $PROJECT_NAME-mysql mysql -u$MYSQL_ROOT_USER -p$MYSQL_ROOT_PASSWORD -e "GRANT ALL ON $MYSQL_DATABASE.* TO '$MYSQL_USER'@'%' IDENTIFIED BY '$MYSQL_PASSWORD';FLUSH PRIVILEGES;"
+#    else
+#      docker container exec -it $PROJECT_NAME-mysql mysql -u$MYSQL_ROOT_USER -p$MYSQL_ROOT_PASSWORD -e "GRANT ALL ON $MYSQL_DATABASE.* TO '$MYSQL_USER'@'%' IDENTIFIED BY '$MYSQL_PASSWORD';FLUSH PRIVILEGES;"
+#    fi
+#fi
 
 # Show the final result
 listString=("$projectUrl")